Move parser to Block class
